requirements.txt
matplotlib
requests
pandas
schedule
python-dotenv
Pillow
imagehash